+++ tests/test_coils.py
@@ -307,6 +307,36 @@ class TestCoilSet:
         )[0]
         np.testing.assert_allclose(B_true, B_approx, rtol=1e-3, atol=1e-10)
 
+    def test_from_symmetry_warnings(self):
+        """Test warning in from_symmetry for symmetry plane and self-intersection."""
+        N = 40
+        # test warning after crossing symmetry plane
+        # tilt coils so they cross the symmetry plane
+        coil = FourierPlanarCoil(normal=[1e-4, 1, 3])
+        coils_list = [coil] + [coil.copy() for i in range(N // 4 - 1)]
+        coils_list_sym = [coil.copy()] + [coil.copy() for i in range(N // 8 - 1)]
+
+        for i, c in enumerate(coils_list[1:]):
+            c.rotate(angle=2 * np.pi / N * (i + 1))
+        for i, c in enumerate(coils_list_sym[1:]):
+            c.rotate(angle=2 * np.pi / N * (i + 1))
+
+        # test just the warning for crossing phi=2pi/NFP plane
+        # crosses plane but since it just a rigid toroidal rotation,
+        # they are not self-intersecting
+        with pytest.warns(UserWarning, match="symmetry"):
+            _ = CoilSet.from_symmetry(coils_list, NFP=4)
+
+        # test the warning for crossing phi=0 plane and for
+        # self-intersecting coils, as two of the coils lie nearly
+        # in the same physical space (intersecting at 2 points) after reflection
+        with pytest.warns(UserWarning) as warninfo:
+            _ = CoilSet.from_symmetry(coils_list_sym, NFP=4, sym=True)
+        assert "phi=0" in str(warninfo[0].message)
+        # "nearly intersecting" warning is given because the grids are finite that
+        # we check the coils on
+        assert "nearly intersecting" in str(warninfo[-1].message)
+
     @pytest.mark.unit
     def test_properties(self):
         """Test getting/setting of CoilSet attributes."""