+++ CHANGELOG.md
@@ -9,8 +9,11 @@ to a ``FourierRZCurve`` object, and ``to_FourierRZCurve`` methods to ``Curve`` c
 coil for each coil in a coilset.
 - Adds the objective `PlasmaCoilsetMinDistance`, which returns the minimum distance to the
 plasma surface for each coil in a coilset.
-- Add method ``calculate_minimum_intercoil_distances`` to ``CoilSet``, which returns the minimum distance to the
+- Add method ``compute_minimum_intercoil_distances`` to ``CoilSet``, which returns the minimum distance to the
 plasma surface for each coil in a coilset.
+- Changes error to warning for ``from_symmetry`` method of ``CoilSet`` when a coil crosses the symmetry plane,
+and adds a check for intersection when the warning is tripped, to allow for valid coilsets which may cross the
+symmetry plane but not be self-intersecting after rotation/reflection.
 
 v0.11.1
 -------